feat: add debugbar library
i intended to use 'patch' instead of 'put' to update a resource, with the intention to avoid unnecessary updates - on fields that haven't changed: the input from 'edit' continued to had the same data initially loaded. In order to achieve that, i made a helper that would do the following: get the formData; filter out values that have changed compared to the model instance that 'edit' view has loaded and values from "special fields" - that aren't stored directly in the database which when null are ignored, such as 'password'. The helper works considering the "Farmer" data, but the change is almost meaningless: already the '$model->update()' only updates changed data and ignore fields not persisted such as 'pass_confirmation'. Still, the library is will remain available to check queries performed.