refactor: unuse "getUpdateData" and change 'update' method from Patch to Post in Farmer routing
the "$model->update", as said in recent commits, already updates only fields that are present in the database and ignore unchanged ones - it avoids unnecessary updates due to sending, for example, <input name="full_name" value="{{$farmer->full_name}}" without change: the form field's data with the same data from the model, and thus the same in the corresponding database record's field. And the difference between using Patch or Put on updating a resource, considering this all, is minimal, so for "enforcing a pattern" - most Laravel applications and developers i know use always 'Put' for updating -, i'm gonna only also use 'Put'.