fix: reorder route definitions to avoid conflicts
considering that routing in laravel implies roughly said the reading of the 'web.php' file from top to bottom in search for a matching route, getting in '/farmers/create' would, if the 'show' route is mentioned before it - '/farmers/{farmer} -, result in the request being sent to not the intended route. To avoid that, the routes got reordered.